% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dps-data.R
\docType{data}
\name{dps}
\alias{dps}
\title{dps genetic distance matrix for Columbia spotted frog (Rana luteiventris)}
\format{A 29 x 29 genetic distance matrix:}
\description{
Subset of data used in Murphy et al., (2010)
}
\references{
Murphy M.A., R. Dezzani, D.S. Pilliod & A.S. Storfer (2010) Landscape genetics of high mountain frog metapopulations. Molecular Ecology 19(17):3634-3649
}
