% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.statistics.R
\name{node.statistics}
\alias{node.statistics}
\title{raster statistics for nodes}
\usage{
node.statistics(x, r, buffer = NULL, stats = c("min", "median", "max"))
}
\arguments{
\item{x}{sp class SpatialPointsDataFrame object}

\item{r}{A rasterLayer, rasterStack or rasterBrick object}

\item{buffer}{Buffer distance, radius in projection units}

\item{stats}{Statistics to calculate. If vectorized, can pass a 
custom statistic function.}
}
\value{
data.frame object of at-node raster values or statistics
}
\description{
returns raster value or statistics 
            (based on specified radius) for node
}
\note{
If no buffer is specified, at-node raster values are returned
}
