% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.gravity.R
\name{plot.gravity}
\alias{plot.gravity}
\title{Plot gravity model}
\usage{
\method{plot}{gravity}(x, type = 1, ...)
}
\arguments{
\item{x}{Object of class gravity}

\item{type}{Type of plot (default 1, model structure I)}

\item{...}{Ignored}
}
\value{
defined plot
}
\description{
Diagnostic plots gravity model with 6 optional plots.
}
\note{
Plot types available: 1 - Model structure I, 2 - Model structure II, 3 - Q-Q Normal - Origin random effects, 4 - Q-Q Normal - Residuals , 5 - Fitted values, 6 - Distribution of observed verses predicted

Depends: nlme, lattice
}
\references{
Murphy, M. A. & J.S. Evans. (in prep). "GenNetIt: gravity analysis in R for landscape genetics"

Murphy M.A., R. Dezzani, D.S. Pilliod & A.S. Storfer (2010) Landscape genetics of high mountain frog metapopulations. Molecular Ecology 19(17):3634-3649
}
\author{
Jeffrey S. Evans  <jeffrey_evans@tnc.org> and Melanie Murphy <melanie.murphy@uwyo.edu>
}
