% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.gravity.R
\name{print.gravity}
\alias{print.gravity}
\title{Print gravity model}
\usage{
\method{print}{gravity}(x, ...)
}
\arguments{
\item{x}{Object of class gravity}

\item{...}{Ignored}
}
\description{
summary method for class "gravity"
}
