% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ralu.site-data.R
\docType{data}
\name{ralu.site}
\alias{ralu.site}
\title{Subset of site-level spatial point data for Columbia spotted frog (Rana luteiventris)}
\format{
An sf POINT object with 31 obs. of  17 variables:
\describe{
\item{SiteName}{Unique site name}
\item{Drainage}{Source drainage}
\item{Basin}{source basin}
\item{Substrate}{Wetland substrate}
\item{NWI}{USFWS NWI Wetland type}
\item{AREA_m2}{Area of wetland}
\item{PERI_m}{Perimeter of wetland}
\item{Depth_m}{Depth of wetland}
\item{TDS}{...}
\item{FISH}{Fish present}
\item{ACB}{...}
\item{AUC}{...}
\item{AUCV}{...}
\item{AUCC}{...}
\item{AUF}{...}
\item{AWOOD}{...}
\item{AUFV}{...}
}
}
\description{
Subset of data used in Murphy et al., (2010)
}
\references{
Murphy M.A., R. Dezzani, D.S. Pilliod & A.S. Storfer (2010) Landscape genetics of high mountain frog metapopulations. Molecular Ecology 19(17):3634-3649
}
