% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.gravity.R
\name{summary.gravity}
\alias{summary.gravity}
\title{Summarizing Gravity Model Fits}
\usage{
\method{summary}{gravity}(object, ...)
}
\arguments{
\item{object}{Object of class gravity}

\item{...}{Ignored}
}
\description{
Summary method for class "gravity".
}
\note{
Summary of lme or lm gravity model, AIC, log likelihood and Root Mean Square Error (RMSE) of observed verses predicted
}
