% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.graph.statistics.R
\name{area.graph.statistics}
\alias{area.graph.statistics}
\title{Statistics for edges (lines) based on a defined scale (area).}
\usage{
area.graph.statistics(...)
}
\arguments{
\item{...}{Parameters to be passed to the modern version of the function}
}
\description{
Samples rasters for each edge and calculates specified 
             statistics for buffer distance
}
\note{
Please note that this function has been deprecated, please use graph.statistics 
      with the buffer argument.
}
