% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.gravity.R
\name{predict.gravity}
\alias{predict.gravity}
\title{Predict gravity model}
\usage{
\method{predict}{gravity}(object, newdata, ...)
}
\arguments{
\item{object}{Object of class gravity}

\item{newdata}{New data, matching model parameters, used for obtaining the predictions}

\item{...}{Arguments passed to predict.lme or predict.lm}
}
\value{
Model predictions
}
\description{
predict method for class "gravity"
}
